+++ contracts/masset/Masset.sol
@@ -186,7 +186,7 @@ contract Masset is IMasset, MassetToken, MassetBasket {
                 require(IERC20(bAsset.addr).transferFrom(msg.sender, address(this), _bassetQuantity[j]), "Basset transfer failed");
                 basket.bassets[indexes[j]].vaultBalance = bAsset.vaultBalance.add(_bassetQuantity[j]);
 
-                uint ratioedBasset = _bassetQuantity[j].mulRatioTruncate(bAsset.ratio);
+                uint256 ratioedBasset = _bassetQuantity[j].mulRatioTruncate(bAsset.ratio);
                 massetQuantity = massetQuantity.add(ratioedBasset);
             }
         }
@@ -256,9 +256,9 @@ contract Masset is IMasset, MassetToken, MassetBasket {
         uint256 massetQuantity = 0;
 
         // Calc MassetQ and update the Vault
-        for(uint i = 0; i < redemptionAssetCount; i++){
+        for(uint256 i = 0; i < redemptionAssetCount; i++){
             if(_bassetQuantities[i] > 0){
-                uint ratioedBasset = _bassetQuantities[i].mulRatioTruncateCeil(bAssets[i].ratio);
+                uint256 ratioedBasset = _bassetQuantities[i].mulRatioTruncateCeil(bAssets[i].ratio);
                 massetQuantity = massetQuantity.add(ratioedBasset);
 
                 // bAsset == bAssets[i] == basket.bassets[indexes[i]]
@@ -276,7 +276,7 @@ contract Masset is IMasset, MassetToken, MassetBasket {
         _burn(msg.sender, massetQuantity);
 
         // Transfer the Bassets to the user
-        for(uint i = 0; i < redemptionAssetCount; i++){
+        for(uint256 i = 0; i < redemptionAssetCount; i++){
             if(_bassetQuantities[i] > 0){
                 require(IERC20(bAssets[i].addr).transfer(_recipient, _bassetQuantities[i]), "Must be successful transfer");
             }
@@ -467,7 +467,7 @@ contract Masset is IMasset, MassetToken, MassetBasket {
     function completeRecol(address _basset, uint256 _unitsUnderCollateralised)
     external
     onlyManager {
-        (bool exists, uint i) = _isAssetInBasket(_basset);
+        (bool exists, uint256 i) = _isAssetInBasket(_basset);
         require(exists, "Basset must exist in Basket");
 
         (, , , , , BassetStatus status) = _getBasset(i);
